code/python_generator/requirements.txt
openai
argparse
pyyaml